PyMethod is internal concept of VirtualMachine
